refactor: remove circular import
